+++ CHANGELOG.md
@@ -2,6 +2,47 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [0.4.0] - 2023/07/24
+
+### Added
+
+- Dir API transfer_to method now receives a filter param in
+  order to filter out files during the transfer.
+  (https://github.com/manoss96/fluke/issues/30)
+
+- There now exists a module *fluke.queues* which contains a
+  number of classes used in order to access message queues
+  provided by various message queue services in the cloud.
+  (https://github.com/manoss96/fluke/issues/31)
+
+- Added a new dependency: ``azure-storage-queue==12.6.0``
+  (https://github.com/manoss96/fluke/issues/31)
+
+### Changed
+
+- *fluke.auth.AzureAuth* no longer has a constructor
+  for creating an *AzureAuth* instance via an Azure
+  service principal. Instead, this is now achieved
+  via an *AzureAuth.from_service_principal* method.
+  (https://github.com/manoss96/fluke/issues/31)
+
+- Classes *fluke.storage.{AWSS3File, AWSS3Dir}*
+  have been renamed to *fluke.storage.{AmazonS3File, AmazonS3Dir}*.
+  (https://github.com/manoss96/fluke/issues/34)
+
+- Updated dependency versions:
+  + azure-identity: 1.12.0 -> 1.13.0
+  + azure-storage-blob: 12.15.0 -> 12.17.0
+  + boto3: 1.26.84 -> 1.28.0
+  + paramiko: 3.0.0 -> 3.2.0
+
+### Fixed
+
+- Fixed issue where type of objects could not be inferred
+  when they were instantiate via a context manager.
+  (https://github.com/manoss96/fluke/issues/35)
+
+
 ## [0.3.0] - 2023/04/16
 
 ### Added
@@ -90,17 +131,17 @@ All notable changes to this project will be documented in this file.
   with the directory that spawned them. In the case of remote files, they
   also share the same client and cache. (https://github.com/manoss96/fluke/issues/16)
 
-- *fluke.storage.AWSS3File* now has a *get_bucket_name* method.
+- *fluke.storage.AmazonS3File* now has a *get_bucket_name* method.
   (https://github.com/manoss96/fluke/issues/16)
    
 ### Changed
 
-- Parameter *path* in *fluke.storage.AWSS3Dir* and *fluke.storage.AzureBlobDir*
+- Parameter *path* in *fluke.storage.AmazonS3Dir* and *fluke.storage.AzureBlobDir*
   constructors now defaults to ``None`` in order to reference the whole
   container and bucket respectively (https://github.com/manoss96/fluke/issues/15)
 
 - Providing a path that starts with a separator, e.g. ``/``, to constructors
-  *fluke.storage.AWSS3Dir* and *fluke.storage.AzureBlobDir* will now throw
+  *fluke.storage.AmazonS3Dir* and *fluke.storage.AzureBlobDir* will now throw
   an ``InvalidPathError`` exception. (https://github.com/manoss96/fluke/issues/15)
 
 - Both *File* and *Dir* API methods *get_metadata* will now return